Products CRUD frontend fully working

+++ backend/controller/products.controller.js
@@ -48,30 +48,54 @@ module.exports = {
   update: (req, res) => {
     let body = JSON.stringify(req.body);
     body = JSON.parse(body);
-
+    console.log(body);
     if (req.file) {
       body.productImg = `http://localhost:8080/${req.file.path.replace(/\\/, '/')}`;
     }
 
-    Products.findByIdAndUpdate(req.params.id, body)
-      .then((products) => {
-        let imgRoute = products.productImg;
-        imgRoute = imgRoute.substring(22);
-        console.log(imgRoute);
+    if (!body.productImg) {
+      Products.findById(req.params.id)
+        .then((products) => {
+          console.log(products);
+          console.log(products.productImg);
+          body.productImg = products.productImg;
+        })
+        .then(Products.findByIdAndUpdate(req.params.id, body)
+          .then((products) => {
+            if (products) {
+              res.status(200).json(products);
+            } else {
+              res.status(404).json({ message: 'Not a valid Id!' });
+            }
+          })
+          .catch((err) => {
+            res.status(500).json({ error: err });
+          }));
+    } else {
+      Products.findByIdAndUpdate(req.params.id, body)
+        .then((products) => {
+          let imgRoute = products.productImg;
+          imgRoute = imgRoute.substring(22);
+          console.log(imgRoute);
 
-        fs.unlink(imgRoute, (err) => {
-          if (err) throw err;
+          fs.exists(imgRoute, (exists) => {
+            if (exists) {
+              fs.unlink(imgRoute, (err) => {
+                if (err) throw err;
+              });
+            }
+          });
+
+          if (products) {
+            res.status(200).json(products);
+          } else {
+            res.status(404).json({ message: 'Not a valid Id!' });
+          }
+        })
+        .catch((err) => {
+          res.status(500).json({ error: err });
         });
-
-        if (products) {
-          res.status(200).json(products);
-        } else {
-          res.status(404).json({ message: 'Not a valid Id!' });
-        }
-      })
-      .catch((err) => {
-        res.status(500).json({ error: err });
-      });
+    }
   },
 
   remove: (req, res) => {
@@ -81,8 +105,12 @@ module.exports = {
         imgRoute = imgRoute.substring(22);
         console.log(imgRoute);
 
-        fs.unlink(imgRoute, (err) => {
-          if (err) throw err;
+        fs.exists(imgRoute, (exists) => {
+          if (exists) {
+            fs.unlink(imgRoute, (err) => {
+              if (err) throw err;
+            });
+          }
         });
 
         if (products) {
